+++ libs/react/shelter-operator/src/lib/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './useAdminShelterProfile';
+export * from './useRoom';
 export * from './useShelterCities';
 export * from './useShelterSpas';
 export * from './useUpdateShelterProfile';
