+++ libs/react/shelter-operator/src/lib/hooks/useRoom/index.ts
@@ -1,0 +1,2 @@
+export type { UseRoomResultType } from './types';
+export { useRoom } from './useRoom';
