+++ libs/react/shelter-operator/src/lib/hooks/useRoom/types.ts
@@ -1,0 +1,3 @@
+import { GetRoomQuery } from '../../components/rooms/api/__generated__/roomQueries.generated';
+
+export type UseRoomResultType = GetRoomQuery['room'];
